requirements.txt
mistralai
pillow
ipython
dotenv